data/pyproject.toml
[tool.poetry]
name = "ragbase"
version = "0.1.0"
description = ""
authors = ["Venelin Valkov <venelin@curiousily.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
langchain-community = "^0.2.6"
flashrank = "^0.2.8"
qdrant-client = "^1.10.1"
pypdfium2 = "^4.30.0"
fastembed = "^0.3.3"
langchain-qdrant = "^0.1.1"
streamlit = "^1.47.1"
python-dotenv = "^1.0.1"
langchain-groq = "^0.1.6"
whoosh = "^2.7.4"
faiss-cpu = "^1.11.0.post1"
huggingface-hub = "^0.33.4"
tabulate = "^0.9.0"
sentence-transformers = "^5.0.0"
pymupdf = "^1.26.3"
python-docx = "^1.2.0"
openai = "^1.97.0"
st-pages = "^1.0.1"
streamlit-mic-recorder = "^0.0.8"
pytube = ">=15.0.0"
openai-whisper = "^20250625"
google-api-python-client = "^2.177.0"
yt-dlp = "^2025.7.21"
deepgram-sdk = "^4.8.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.1"
jupyterlab = "^4.2.3"
watchdog = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
